[refactor] #133 socialLogin Live, Test 분리
